fix: use int64_t and overflow-checked arithmetic in shape inference
Promote MathOpTwoIntegers() from int to int64_t and guard Add and Mul
with checked_add_overflow / checked_mul_overflow so that data
propagation through arithmetic nodes raises InferenceError instead of
silently producing a wrapped result.

Add checked_add_overflow to safe_math.h, following the same pattern
as the existing checked_mul_overflow.

Fix the Concat shape inference accumulator: change total_length from
int to int64_t and remove the static_cast<int> truncation on each
input dimension, eliminating a truncation from int64 to 32-bit int.

+++ onnx/defs/math/utils.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -52,7 +53,7 @@ void QLinearMatMulShapeInference(ONNX_NAMESPACE::InferenceContext& ctx);
 
 const char* QLinearMatMulDoc();
 
-int MathOpTwoIntegers(const std::string& op_type, int a, int b);
+int64_t MathOpTwoIntegers(const std::string& op_type, int64_t a, int64_t b);
 
 } // namespace utils
 } // namespace math
